port commands over to new system
