+++ Firmware/src/serialhandler/cli/help.cpp
@@ -1,11 +1,11 @@
 #include "global.h"
 
-void initHelp(){
+void initHelp() {
 	Command *cmd = new Command("help", "Print list of available commands");
 	cmd->setExecuteFunction([](std::map<string, RuntimeArg> &args, Command *cmd) {
-		string response = "Available commands:\n";
+		string response = "Available commands:";
 		for (Command *c : Command::cliCommands) {
-			response += "  " + c->name + " - " + c->description + "\n";
+			response += "\n  " + c->name + " - " + c->description;
 		}
 		cmd->print(response.c_str());
 		return false;
