+++ Firmware/src/serialhandler/cli/save.cpp
@@ -1,0 +1,12 @@
+#include "global.h"
+
+void initSave() {
+	Command *cmd = new Command("save", "Save current settings to file");
+	cmd->setExecuteFunction([](std::map<string, RuntimeArg> &args, Command *cmd) {
+		closeSettingsFile();
+		openSettingsFile();
+		cmd->print("Settings saved");
+		return false;
+	});
+	Command::cliCommands.push_back(cmd);
+}
